homedashboard_django/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
amqp==5.3.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
attrs==25.3.0
bcrypt==4.3.0
billiard==4.2.1
cachetools==5.5.2
celery==5.4.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap4==2022.1
cron-descriptor==1.4.5
cryptography==41.0.5
defusedxml==0.8.0rc2
discord.py==2.5.2
Django==4.2
django-celery-beat==2.7.0
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-environ==0.10.0
django-filter==23.1
django-json-widget==2.0.1
django-rest-framework==0.1.0
django-templated-mail==1.1.1
django-timezone-field==7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
Faker==20.0.0
frozenlist==1.6.0
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.68.0
httplib2==0.22.0
idna==3.4
kombu==5.4.2
Markdown==3.4.3
multidict==6.4.4
oauthlib==3.2.2
paramiko==3.5.1
Pillow==9.5.0
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
PyJWT==2.8.0
PyNaCl==1.5.0
pyparsing==3.2.1
python-crontab==3.2.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
setuptools==75.8.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.0
